Fix for uptime and downtime constraints
Fixed several issues:
(1) When a generator is turned on/off, rather than having to stay in that state for the next n hours, it need only be in that state for any n hours across the optimization.
(2) A generator could not be turned on/off within a margin of minDowntime/minUptime hours from the end of the optimization.
(3) If a generator was turned off/on within that margin of hour 25, then in the subsequent optimization (in which the old hour 25 becomes hour 1), that generator must remain in that state for the full optimization.